+++ docs/RELEASE-NOTES.rst
@@ -9,10 +9,12 @@ Added Functionality
 **What's new:**
     * Multi Cluster
     * CRD
+      * Support for mix of k8s Secret and bigip reference in TLSProfile.
 
 Bug Fixes
 ````````````
 * `Issue 3230 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/3230>`_: CRD multicluster configuration triggers Raw response from Big-IP: map[code:422 declarationFullId: message:declaration has duplicate values in rules].
+* `Issue 3239 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/3239>`_: Support for mix of k8s Secret and bigip reference in TLSProfile.
 
 2.15.1
 -------------